Update RasterizeRDD tests
